# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal AI agent management API — multi-agent routing, LLM providers, tools, approvals, and more.",
        "title": "Denkeeper API",
        "contact": {
            "name": "Denkeeper",
            "url": "https://github.com/Temikus/denkeeper"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/kv/{agent}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all key-value entries for the specified agent, optionally filtered by a key prefix.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "List KV entries for an agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent name",
                        "name": "agent",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter keys by prefix",
                        "name": "prefix",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object with entries array containing key, value, created_at, updated_at, and optional expires_at",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Agent not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "KV store not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/kv/{agent}/{key}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the value of a single key for the specified agent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "Get a KV entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent name",
                        "name": "agent",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key name",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object with key and value fields",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Agent or key not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "KV store not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates a key-value pair for the specified agent. Optionally accepts a TTL duration string (e.g. \"1h\", \"30m\").",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kv"
                ],
                "summary": "Set a KV entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent name",
                        "name": "agent",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key name",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Value and optional TTL",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object with key and value fields",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON body or TTL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Agent not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "KV store not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a single key-value pair for the specified agent.",
                "tags": [
                    "kv"
                ],
                "summary": "Delete a KV entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent name",
                        "name": "agent",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Key name",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Key deleted"
                    },
                    "404": {
                        "description": "Agent not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "KV store not configured",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "API key prefixed with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}